+++ src/topology/gnunet-daemon-topology.c
@@ -841,7 +841,8 @@ static int
 address_iterator (void *cls,
 		  const char *tname,
 		  struct GNUNET_TIME_Absolute expiration,
-		  const void *addr, size_t addrlen)
+		  const void *addr, 
+		  uint16_t addrlen)
 {
   int *flag = cls;
   *flag = GNUNET_YES;
